Moves all other tests, except for the stack related ones, from ubpf_vm into ubpf_unified.
