+++ tests/ubpf_vm.rs
@@ -20,242 +20,8 @@
 // These are unit tests for the eBPF interpreter.
 
 extern crate solana_rbpf;
-mod common;
 
-use common::{TCP_SACK_ASM, TCP_SACK_MATCH, TCP_SACK_NOMATCH};
 use solana_rbpf::{assembler::assemble, syscalls, user_error::UserError, vm::EbpfVm};
-
-#[test]
-fn test_vm_alu64_arith() {
-    let prog = assemble(
-        "
-        mov r0, 0
-        mov r1, 1
-        mov r2, 2
-        mov r3, 3
-        mov r4, 4
-        mov r5, 5
-        mov r6, 6
-        mov r7, 7
-        mov r8, 8
-        mov r9, 9
-        add r0, 23
-        add r0, r7
-        sub r0, 13
-        sub r0, r1
-        mul r0, 7
-        mul r0, r3
-        div r0, 2
-        div r0, r4
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x2a);
-}
-
-#[test]
-fn test_vm_alu64_bit() {
-    let prog = assemble(
-        "
-        mov r0, 0
-        mov r1, 1
-        mov r2, 2
-        mov r3, 3
-        mov r4, 4
-        mov r5, 5
-        mov r6, 6
-        mov r7, 7
-        mov r8, 8
-        or r0, r5
-        or r0, 0xa0
-        and r0, 0xa3
-        mov r9, 0x91
-        and r0, r9
-        lsh r0, 32
-        lsh r0, 22
-        lsh r0, r8
-        rsh r0, 32
-        rsh r0, 19
-        rsh r0, r7
-        xor r0, 0x03
-        xor r0, r2
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x11);
-}
-
-#[test]
-fn test_vm_alu_arith() {
-    let prog = assemble(
-        "
-        mov32 r0, 0
-        mov32 r1, 1
-        mov32 r2, 2
-        mov32 r3, 3
-        mov32 r4, 4
-        mov32 r5, 5
-        mov32 r6, 6
-        mov32 r7, 7
-        mov32 r8, 8
-        mov32 r9, 9
-        add32 r0, 23
-        add32 r0, r7
-        sub32 r0, 13
-        sub32 r0, r1
-        mul32 r0, 7
-        mul32 r0, r3
-        div32 r0, 2
-        div32 r0, r4
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x2a);
-}
-
-#[test]
-fn test_vm_arsh32_high_shift() {
-    let prog = assemble(
-        "
-        mov r0, 8
-        lddw r1, 0x100000001
-        arsh32 r0, r1
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x4);
-}
-
-#[test]
-fn test_vm_arsh() {
-    let prog = assemble(
-        "
-        mov32 r0, 0xf8
-        lsh32 r0, 28
-        arsh32 r0, 16
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xffff8000);
-}
-
-#[test]
-fn test_vm_arsh64() {
-    let prog = assemble(
-        "
-        mov32 r0, 1
-        lsh r0, 63
-        arsh r0, 55
-        mov32 r1, 5
-        arsh r0, r1
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xfffffffffffffff8);
-}
-
-#[test]
-fn test_vm_arsh_reg() {
-    let prog = assemble(
-        "
-        mov32 r0, 0xf8
-        mov32 r1, 16
-        lsh32 r0, 28
-        arsh32 r0, r1
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xffff8000);
-}
-
-#[test]
-fn test_vm_be16() {
-    let prog = assemble(
-        "
-        ldxh r0, [r1]
-        be16 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x11, 0x22];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1122);
-}
-
-#[test]
-fn test_vm_be16_high() {
-    let prog = assemble(
-        "
-        ldxdw r0, [r1]
-        be16 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1122);
-}
-
-#[test]
-fn test_vm_be32() {
-    let prog = assemble(
-        "
-        ldxw r0, [r1]
-        be32 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x11, 0x22, 0x33, 0x44];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x11223344);
-}
-
-#[test]
-fn test_vm_be32_high() {
-    let prog = assemble(
-        "
-        ldxdw r0, [r1]
-        be32 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x11223344);
-}
-
-#[test]
-fn test_vm_be64() {
-    let prog = assemble(
-        "
-        ldxdw r0, [r1]
-        be64 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1122334455667788);
-}
 
 // TODO: syscalls::trash_registers needs asm!().
 // Try this again once asm!() is available in stable.
@@ -285,25 +51,6 @@ fn test_vm_be64() {
 //fn test_vm_err_call_bad_imm() {
 //}
 
-#[test]
-#[should_panic(expected = "UnresolvedSymbol(\"Unknown\", 34, 40)")]
-fn test_vm_err_call_unreg() {
-    let prog = assemble(
-        "
-        mov r1, 1
-        mov r2, 2
-        mov r3, 3
-        mov r4, 4
-        mov r5, 5
-        call 63
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    vm.execute_program().unwrap();
-}
-
 // With the introduction of call frames there may be stack regions
 // above or below the current stack, to test out of bounds we have to
 // try significantly further away
@@ -319,375 +66,6 @@ fn test_vm_err_stack_out_of_bound() {
     let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
     let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
     vm.execute_program().unwrap();
-}
-
-#[test]
-fn test_vm_ldxb_all() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        ldxb r9, [r0+0]
-        lsh r9, 0
-        ldxb r8, [r0+1]
-        lsh r8, 4
-        ldxb r7, [r0+2]
-        lsh r7, 8
-        ldxb r6, [r0+3]
-        lsh r6, 12
-        ldxb r5, [r0+4]
-        lsh r5, 16
-        ldxb r4, [r0+5]
-        lsh r4, 20
-        ldxb r3, [r0+6]
-        lsh r3, 24
-        ldxb r2, [r0+7]
-        lsh r2, 28
-        ldxb r1, [r0+8]
-        lsh r1, 32
-        ldxb r0, [r0+9]
-        lsh r0, 36
-        or r0, r1
-        or r0, r2
-        or r0, r3
-        or r0, r4
-        or r0, r5
-        or r0, r6
-        or r0, r7
-        or r0, r8
-        or r0, r9
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, //
-        0x08, 0x09, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x9876543210);
-}
-
-#[test]
-fn test_vm_ldxh_all() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        ldxh r9, [r0+0]
-        be16 r9
-        lsh r9, 0
-        ldxh r8, [r0+2]
-        be16 r8
-        lsh r8, 4
-        ldxh r7, [r0+4]
-        be16 r7
-        lsh r7, 8
-        ldxh r6, [r0+6]
-        be16 r6
-        lsh r6, 12
-        ldxh r5, [r0+8]
-        be16 r5
-        lsh r5, 16
-        ldxh r4, [r0+10]
-        be16 r4
-        lsh r4, 20
-        ldxh r3, [r0+12]
-        be16 r3
-        lsh r3, 24
-        ldxh r2, [r0+14]
-        be16 r2
-        lsh r2, 28
-        ldxh r1, [r0+16]
-        be16 r1
-        lsh r1, 32
-        ldxh r0, [r0+18]
-        be16 r0
-        lsh r0, 36
-        or r0, r1
-        or r0, r2
-        or r0, r3
-        or r0, r4
-        or r0, r5
-        or r0, r6
-        or r0, r7
-        or r0, r8
-        or r0, r9
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x00, 0x00, 0x00, 0x01, 0x00, 0x02, 0x00, 0x03, //
-        0x00, 0x04, 0x00, 0x05, 0x00, 0x06, 0x00, 0x07, //
-        0x00, 0x08, 0x00, 0x09, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x9876543210);
-}
-
-#[test]
-fn test_vm_ldxh_all2() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        ldxh r9, [r0+0]
-        be16 r9
-        ldxh r8, [r0+2]
-        be16 r8
-        ldxh r7, [r0+4]
-        be16 r7
-        ldxh r6, [r0+6]
-        be16 r6
-        ldxh r5, [r0+8]
-        be16 r5
-        ldxh r4, [r0+10]
-        be16 r4
-        ldxh r3, [r0+12]
-        be16 r3
-        ldxh r2, [r0+14]
-        be16 r2
-        ldxh r1, [r0+16]
-        be16 r1
-        ldxh r0, [r0+18]
-        be16 r0
-        or r0, r1
-        or r0, r2
-        or r0, r3
-        or r0, r4
-        or r0, r5
-        or r0, r6
-        or r0, r7
-        or r0, r8
-        or r0, r9
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x00, 0x01, 0x00, 0x02, 0x00, 0x04, 0x00, 0x08, //
-        0x00, 0x10, 0x00, 0x20, 0x00, 0x40, 0x00, 0x80, //
-        0x01, 0x00, 0x02, 0x00, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x3ff);
-}
-
-#[test]
-fn test_vm_ldxw_all() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        ldxw r9, [r0+0]
-        be32 r9
-        ldxw r8, [r0+4]
-        be32 r8
-        ldxw r7, [r0+8]
-        be32 r7
-        ldxw r6, [r0+12]
-        be32 r6
-        ldxw r5, [r0+16]
-        be32 r5
-        ldxw r4, [r0+20]
-        be32 r4
-        ldxw r3, [r0+24]
-        be32 r3
-        ldxw r2, [r0+28]
-        be32 r2
-        ldxw r1, [r0+32]
-        be32 r1
-        ldxw r0, [r0+36]
-        be32 r0
-        or r0, r1
-        or r0, r2
-        or r0, r3
-        or r0, r4
-        or r0, r5
-        or r0, r6
-        or r0, r7
-        or r0, r8
-        or r0, r9
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, //
-        0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x08, //
-        0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, //
-        0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x08, 0x00, //
-        0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x030f0f);
-}
-
-#[test]
-fn test_vm_le16() {
-    let prog = assemble(
-        "
-        ldxh r0, [r1]
-        le16 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0x22, 0x11];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1122);
-}
-
-#[test]
-fn test_vm_le32() {
-    let prog = assemble(
-        "
-        ldxw r0, [r1]
-        le32 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x44, 0x33, 0x22, 0x11, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x11223344);
-}
-
-#[test]
-fn test_vm_le64() {
-    let prog = assemble(
-        "
-        ldxdw r0, [r1]
-        le64 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x88, 0x77, 0x66, 0x55, 0x44, 0x33, 0x22, 0x11, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1122334455667788);
-}
-
-#[test]
-fn test_vm_lsh_reg() {
-    let prog = assemble(
-        "
-        mov r0, 0x1
-        mov r7, 4
-        lsh r0, r7
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x10);
-}
-
-#[test]
-fn test_vm_mul_loop() {
-    let prog = assemble(
-        "
-        mov r0, 0x7
-        add r1, 0xa
-        lsh r1, 0x20
-        rsh r1, 0x20
-        jeq r1, 0x0, +4
-        mov r0, 0x7
-        mul r0, 0x7
-        add r1, -1
-        jne r1, 0x0, -3
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x75db9c97);
-}
-
-#[test]
-fn test_vm_neg64() {
-    let prog = assemble(
-        "
-        mov32 r0, 2
-        neg r0
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xfffffffffffffffe);
-}
-
-#[test]
-fn test_vm_neg() {
-    let prog = assemble(
-        "
-        mov32 r0, 2
-        neg32 r0
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xfffffffe);
-}
-
-#[test]
-fn test_vm_prime() {
-    let prog = assemble(
-        "
-        mov r1, 67
-        mov r0, 0x1
-        mov r2, 0x2
-        jgt r1, 0x2, +4
-        ja +10
-        add r2, 0x1
-        mov r0, 0x1
-        jge r2, r1, +7
-        mov r3, r1
-        div r3, r2
-        mul r3, r2
-        mov r4, r1
-        sub r4, r3
-        mov r0, 0x0
-        jne r4, 0x0, -10
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1);
-}
-
-#[test]
-fn test_vm_rhs32() {
-    let prog = assemble(
-        "
-        xor r0, r0
-        sub r0, 1
-        rsh32 r0, 8
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x00ffffff);
-}
-
-#[test]
-fn test_vm_rsh_reg() {
-    let prog = assemble(
-        "
-        mov r0, 0x10
-        mov r7, 4
-        rsh r0, r7
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1);
 }
 
 #[test]
@@ -776,263 +154,5 @@ fn test_vm_string_stack() {
     let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
     let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
     vm.register_syscall(4, syscalls::strcmp).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x0);
-}
-
-#[test]
-fn test_vm_stxb_all() {
-    let prog = assemble(
-        "
-        mov r0, 0xf0
-        mov r2, 0xf2
-        mov r3, 0xf3
-        mov r4, 0xf4
-        mov r5, 0xf5
-        mov r6, 0xf6
-        mov r7, 0xf7
-        mov r8, 0xf8
-        stxb [r1], r0
-        stxb [r1+1], r2
-        stxb [r1+2], r3
-        stxb [r1+3], r4
-        stxb [r1+4], r5
-        stxb [r1+5], r6
-        stxb [r1+6], r7
-        stxb [r1+7], r8
-        ldxdw r0, [r1]
-        be64 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xf0f2f3f4f5f6f7f8);
-}
-
-#[test]
-fn test_vm_stxb_all2() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        mov r1, 0xf1
-        mov r9, 0xf9
-        stxb [r0], r1
-        stxb [r0+1], r9
-        ldxh r0, [r0]
-        be16 r0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [0xff, 0xff];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0xf1f9);
-}
-
-#[test]
-fn test_vm_stxb_chain() {
-    let prog = assemble(
-        "
-        mov r0, r1
-        ldxb r9, [r0+0]
-        stxb [r0+1], r9
-        ldxb r8, [r0+1]
-        stxb [r0+2], r8
-        ldxb r7, [r0+2]
-        stxb [r0+3], r7
-        ldxb r6, [r0+3]
-        stxb [r0+4], r6
-        ldxb r5, [r0+4]
-        stxb [r0+5], r5
-        ldxb r4, [r0+5]
-        stxb [r0+6], r4
-        ldxb r3, [r0+6]
-        stxb [r0+7], r3
-        ldxb r2, [r0+7]
-        stxb [r0+8], r2
-        ldxb r1, [r0+8]
-        stxb [r0+9], r1
-        ldxb r0, [r0+9]
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x2a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-        0x00, 0x00, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x2a);
-}
-
-#[test]
-fn test_vm_subnet() {
-    let prog = assemble(
-        "
-        mov r2, 0xe
-        ldxh r3, [r1+12]
-        jne r3, 0x81, +2
-        mov r2, 0x12
-        ldxh r3, [r1+16]
-        and r3, 0xffff
-        jne r3, 0x8, +5
-        add r1, r2
-        mov r0, 0x1
-        ldxw r1, [r1+16]
-        and r1, 0xffffff
-        jeq r1, 0x1a8c0, +1
-        mov r0, 0x0
-        exit",
-    )
-    .unwrap();
-    let mem = &mut [
-        0x00, 0x00, 0xc0, 0x9f, 0xa0, 0x97, 0x00, 0xa0, //
-        0xcc, 0x3b, 0xbf, 0xfa, 0x08, 0x00, 0x45, 0x10, //
-        0x00, 0x3c, 0x46, 0x3c, 0x40, 0x00, 0x40, 0x06, //
-        0x73, 0x1c, 0xc0, 0xa8, 0x01, 0x02, 0xc0, 0xa8, //
-        0x01, 0x01, 0x06, 0x0e, 0x00, 0x17, 0x99, 0xc5, //
-        0xa0, 0xec, 0x00, 0x00, 0x00, 0x00, 0xa0, 0x02, //
-        0x7d, 0x78, 0xe0, 0xa3, 0x00, 0x00, 0x02, 0x04, //
-        0x05, 0xb4, 0x04, 0x02, 0x08, 0x0a, 0x00, 0x9c, //
-        0x27, 0x24, 0x00, 0x00, 0x00, 0x00, 0x01, 0x03, //
-        0x03, 0x00, //
-    ];
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1);
-}
-
-const PROG_TCP_PORT_80: [u8; 152] = [
-    0x71, 0x12, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x71, 0x13, 0x0d, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x67, 0x03, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, //
-    0x4f, 0x23, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0xb7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x55, 0x03, 0x0c, 0x00, 0x08, 0x00, 0x00, 0x00, //
-    0x71, 0x12, 0x17, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x55, 0x02, 0x0a, 0x00, 0x06, 0x00, 0x00, 0x00, //
-    0x71, 0x12, 0x0e, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x07, 0x01, 0x00, 0x00, 0x0e, 0x00, 0x00, 0x00, //
-    0x57, 0x02, 0x00, 0x00, 0x0f, 0x00, 0x00, 0x00, //
-    0x67, 0x02, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, //
-    0x0f, 0x21, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x69, 0x12, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x15, 0x02, 0x02, 0x00, 0x00, 0x50, 0x00, 0x00, //
-    0x69, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-    0x55, 0x01, 0x01, 0x00, 0x00, 0x50, 0x00, 0x00, //
-    0xb7, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, //
-    0x95, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, //
-];
-
-#[test]
-fn test_vm_tcp_port80_match() {
-    let mem = &mut [
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x00, 0x06, //
-        0x07, 0x08, 0x09, 0x0a, 0x08, 0x00, 0x45, 0x00, //
-        0x00, 0x56, 0x00, 0x01, 0x00, 0x00, 0x40, 0x06, //
-        0xf9, 0x4d, 0xc0, 0xa8, 0x00, 0x01, 0xc0, 0xa8, //
-        0x00, 0x02, 0x27, 0x10, 0x00, 0x50, 0x00, 0x00, //
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x02, //
-        0x20, 0x00, 0xc5, 0x18, 0x00, 0x00, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, //
-    ];
-    let prog = &PROG_TCP_PORT_80;
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1);
-}
-
-#[test]
-fn test_vm_tcp_port80_nomatch() {
-    let mem = &mut [
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x00, 0x06, //
-        0x07, 0x08, 0x09, 0x0a, 0x08, 0x00, 0x45, 0x00, //
-        0x00, 0x56, 0x00, 0x01, 0x00, 0x00, 0x40, 0x06, //
-        0xf9, 0x4d, 0xc0, 0xa8, 0x00, 0x01, 0xc0, 0xa8, //
-        0x00, 0x02, 0x00, 0x16, 0x27, 0x10, 0x00, 0x00, //
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x51, 0x02, //
-        0x20, 0x00, 0xc5, 0x18, 0x00, 0x00, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, //
-    ];
-    let prog = &PROG_TCP_PORT_80;
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x0);
-}
-
-#[test]
-fn test_vm_tcp_port80_nomatch_ethertype() {
-    let mem = &mut [
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x00, 0x06, //
-        0x07, 0x08, 0x09, 0x0a, 0x08, 0x01, 0x45, 0x00, //
-        0x00, 0x56, 0x00, 0x01, 0x00, 0x00, 0x40, 0x06, //
-        0xf9, 0x4d, 0xc0, 0xa8, 0x00, 0x01, 0xc0, 0xa8, //
-        0x00, 0x02, 0x27, 0x10, 0x00, 0x50, 0x00, 0x00, //
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x02, //
-        0x20, 0x00, 0xc5, 0x18, 0x00, 0x00, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, //
-    ];
-    let prog = &PROG_TCP_PORT_80;
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x0);
-}
-
-#[test]
-fn test_vm_tcp_port80_nomatch_proto() {
-    let mem = &mut [
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x00, 0x06, //
-        0x07, 0x08, 0x09, 0x0a, 0x08, 0x00, 0x45, 0x00, //
-        0x00, 0x56, 0x00, 0x01, 0x00, 0x00, 0x40, 0x11, //
-        0xf9, 0x4d, 0xc0, 0xa8, 0x00, 0x01, 0xc0, 0xa8, //
-        0x00, 0x02, 0x27, 0x10, 0x00, 0x50, 0x00, 0x00, //
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x02, //
-        0x20, 0x00, 0xc5, 0x18, 0x00, 0x00, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, //
-        0x44, 0x44, 0x44, 0x44, //
-    ];
-    let prog = &PROG_TCP_PORT_80;
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem, &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x0);
-}
-
-#[test]
-fn test_vm_tcp_sack_match() {
-    let mut mem = TCP_SACK_MATCH.to_vec();
-    let prog = assemble(TCP_SACK_ASM).unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem.as_mut_slice(), &[]).unwrap();
-    assert_eq!(vm.execute_program().unwrap(), 0x1);
-}
-
-#[test]
-fn test_vm_tcp_sack_nomatch() {
-    let mut mem = TCP_SACK_NOMATCH.to_vec();
-    let prog = assemble(TCP_SACK_ASM).unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), mem.as_mut_slice(), &[]).unwrap();
     assert_eq!(vm.execute_program().unwrap(), 0x0);
 }
